Update stylesheet_instantSearch.css
Any time a 0 in padding or margins, putting any unit with the 0 is confusing to the browser and slows the process.  You may not get an error or warning but, the syntax is incorrect.

Lines 18 and 21 might "bust the box" unless you add the box-sizing:border-box;    https://css-tricks.com/almanac/properties/b/box-sizing/  Note that this will not work with old versions of IE.  Specifically, versions 6 and 7 which were EOL in 2010 before Windows XP

Lines 27-30 are not validating for some reason.  It may be just the fact that flex has a problem with older versions of IE.  I will know better when I load this mod for trial.  https://css-tricks.com/snippets/css/a-guide-to-flexbox/#flexbox-properties