Change search for short terms
For more consistency it's better to search short terms on the entire
name/model as well, and not only at the beginning.
The position of the word in the result is still taken into account
later anyway (i.e. words that occur first have a higher weight
on the results list).